+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Fix incorrectly returning -1 when there is at least one valid range
+  * perf: remove internal function
 
 1.0.3 / 2015-10-29
 ==================